heddle#184: extract per-language LanguageRules trait in semantic merge driver
Refactor candidate A from the heddle#133 audit. Pulls per-language behaviour
in `crates/semantic/src/merge_driver/items.rs` behind a `LanguageRules` trait
in a new `language_rules.rs` module. Six unit-struct implementations
(`RustRules`, `PythonRules`, `JsTsRules`, `GoRules`, `CppRules`, `JavaRules`)
cover the four axes the audit identified: classify_node, leading_metadata_kinds,
signature_hash, extra_scope. Zero behavioural change targeted; the 198 existing
semantic-crate tests pass unchanged. C++'s declarator/qualified-name machinery
stays inside `CppRules` — generalising it is the deferred TW4 (heddle#188), not
this issue.

Ten discipline-guard tests added: per-language snapshots of
`leading_metadata_kinds`, an exhaustive-match guard so adding a new `Language`
variant fails compilation if rules_for() is not extended, and a JS/TS
shared-rules check.
